+++ apps/rslib-module/src/stories/index.stories.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Counter } from 'rslib-module';
+
+const Component = () => <Counter />;
+
+export default {
+  title: 'App Component',
+  component: Component,
+};
+
+export const Primary = {};