+++ packages/managers/src/PKGJsonManager.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import finder from 'find-pkg';
 import fs from 'fs';
-import { MFModuleType } from '@module-federation/sdk';
+import { MFModuleType, logger } from '@module-federation/sdk';
 
 export class PKGJsonManager {
   private _pkg?: Record<string, any>;
@@ -27,7 +27,7 @@ export class PKGJsonManager {
         this._pkg = pkg;
         return pkg;
       } catch (err) {
-        console.error(err);
+        logger.error(err);
         return {};
       }
     }