+++ packages/nextjs-mf/src/plugins/container/runtimePlugin.ts
@@ -7,11 +7,10 @@ export default function (): FederationRuntimePlugin {
       url: string;
       attrs?: Record<string, any>;
     }) {
-      // Updated type
-      var url = args.url;
-      var attrs = args.attrs;
+      const url = args.url;
+      const attrs = args.attrs;
       if (typeof window !== 'undefined') {
-        var script = document.createElement('script');
+        const script = document.createElement('script');
         script.src = url;
         script.async = true;
         delete attrs?.['crossorigin'];
@@ -26,20 +25,21 @@ export default function (): FederationRuntimePlugin {
       from: string;
       origin: any;
     }) {
-      var id = args.id;
-      var error = args.error;
-      var from = args.from;
+      const id = args.id;
+      const error = args.error;
+      const from = args.from;
+      //@ts-ignore
+      globalThis.moduleGraphDirty = true;
       console.error(id, 'offline');
-      var pg = function () {
+      const pg = function () {
         console.error(id, 'offline', error);
         return null;
       };
 
       (pg as any).getInitialProps = function (ctx: any) {
-        // Type assertion to add getInitialProps
         return {};
       };
-      var mod;
+      let mod;
       if (from === 'build') {
         mod = function () {
           return {
@@ -70,15 +70,16 @@ export default function (): FederationRuntimePlugin {
         return args;
       }
 
-      var moduleCache = args.origin.moduleCache;
-      var name = args.origin.name;
-      var gs;
+      const moduleCache = args.origin.moduleCache;
+      const name = args.origin.name;
+      let gs;
       try {
         gs = new Function('return globalThis')();
       } catch (e) {
         gs = globalThis; // fallback for browsers without 'unsafe-eval' CSP policy enabled
       }
-      var attachedRemote = gs[name];
+      //@ts-ignore
+      const attachedRemote = gs[name];
       if (attachedRemote) {
         moduleCache.set(name, attachedRemote);
       }
@@ -89,10 +90,10 @@ export default function (): FederationRuntimePlugin {
       return args;
     },
     beforeRequest: function (args: any) {
-      var options = args.options;
-      var id = args.id;
-      var remoteName = id.split('/').shift();
-      var remote = options.remotes.find(function (remote: any) {
+      const options = args.options;
+      const id = args.id;
+      const remoteName = id.split('/').shift();
+      const remote = options.remotes.find(function (remote: any) {
         return remote.name === remoteName;
       });
       if (!remote) return args;
@@ -106,41 +107,41 @@ export default function (): FederationRuntimePlugin {
       return args;
     },
     onLoad: function (args: any) {
-      var exposeModuleFactory = args.exposeModuleFactory;
-      var exposeModule = args.exposeModule;
-      var id = args.id;
-      var moduleOrFactory = exposeModuleFactory || exposeModule;
-      if (!moduleOrFactory) return args; // Ensure moduleOrFactory is defined
+      const exposeModuleFactory = args.exposeModuleFactory;
+      const exposeModule = args.exposeModule;
+      const id = args.id;
+      const moduleOrFactory = exposeModuleFactory || exposeModule;
+      if (!moduleOrFactory) return args;
 
       if (typeof window === 'undefined') {
-        var exposedModuleExports: any;
+        let exposedModuleExports: any;
         try {
           exposedModuleExports = moduleOrFactory();
         } catch (e) {
           exposedModuleExports = moduleOrFactory;
         }
 
-        var handler: ProxyHandler<any> = {
+        const handler: ProxyHandler<any> = {
           get: function (target, prop, receiver) {
-            // Check if accessing a static property of the function itself
             if (
               target === exposedModuleExports &&
               typeof exposedModuleExports[prop] === 'function'
             ) {
               return function (this: unknown) {
                 globalThis.usedChunks.add(id);
+                //eslint-disable-next-line
                 return exposedModuleExports[prop].apply(this, arguments);
               };
             }
 
-            var originalMethod = target[prop];
+            const originalMethod = target[prop];
             if (typeof originalMethod === 'function') {
-              var proxiedFunction = function (this: unknown) {
+              const proxiedFunction = function (this: unknown) {
                 globalThis.usedChunks.add(id);
+                //eslint-disable-next-line
                 return originalMethod.apply(this, arguments);
               };
 
-              // Copy all enumerable properties from the original method to the proxied function
               Object.keys(originalMethod).forEach(function (prop) {
                 Object.defineProperty(proxiedFunction, prop, {
                   value: originalMethod[prop],
@@ -158,12 +159,9 @@ export default function (): FederationRuntimePlugin {
         };
 
         if (typeof exposedModuleExports === 'function') {
-          // If the module export is a function, we create a proxy that can handle both its
-          // call (as a function) and access to its properties (including static methods).
           exposedModuleExports = new Proxy(exposedModuleExports, handler);
 
-          // Proxy static properties specifically
-          var staticProps = Object.getOwnPropertyNames(exposedModuleExports);
+          const staticProps = Object.getOwnPropertyNames(exposedModuleExports);
           staticProps.forEach(function (prop) {
             if (typeof exposedModuleExports[prop] === 'function') {
               exposedModuleExports[prop] = new Proxy(
@@ -176,7 +174,6 @@ export default function (): FederationRuntimePlugin {
             return exposedModuleExports;
           };
         } else {
-          // For objects, just wrap the exported object itself
           exposedModuleExports = new Proxy(exposedModuleExports, handler);
         }
 
@@ -194,12 +191,12 @@ export default function (): FederationRuntimePlugin {
       ) {
         return args;
       }
-      var shareScopeMap = args.shareScopeMap;
-      var scope = args.scope;
-      var pkgName = args.pkgName;
-      var version = args.version;
-      var GlobalFederation = args.GlobalFederation;
-      var host = GlobalFederation['__INSTANCES__'][0];
+      const shareScopeMap = args.shareScopeMap;
+      const scope = args.scope;
+      const pkgName = args.pkgName;
+      const version = args.version;
+      const GlobalFederation = args.GlobalFederation;
+      const host = GlobalFederation['__INSTANCES__'][0];
       if (!host) {
         return args;
       }
@@ -207,10 +204,9 @@ export default function (): FederationRuntimePlugin {
       if (!host.options.shared[pkgName]) {
         return args;
       }
-      //handle react host next remote, disable resolving when not next host
       args.resolver = function () {
         shareScopeMap[scope][pkgName][version] =
-          host.options.shared[pkgName][0]; // replace local share scope manually with desired module
+          host.options.shared[pkgName][0];
         return shareScopeMap[scope][pkgName][version];
       };
       return args;