+++ packages/node/src/index.ts
@@ -1,6 +1,7 @@
 export { default as StreamingTargetPlugin } from './plugins/StreamingTargetPlugin';
 export { default as NodeFederationPlugin } from './plugins/NodeFederationPlugin';
 export { default as UniversalFederationPlugin } from './plugins/UniversalFederationPlugin';
+//@ts-ignore
 export { default as ChunkCorrelationPlugin } from './plugins/ChunkCorrelationPlugin';
 export { default as RemotePublicPathPlugin } from './plugins/RemotePublicPathRuntimeModule';
 export { default as EntryChunkTrackerPlugin } from './plugins/EntryChunkTrackerPlugin';