+++ packages/runtime/src/module/index.ts
@@ -1,4 +1,4 @@
-import { getFMId, assert } from '../utils';
+import { getFMId, assert, logger } from '../utils';
 import { safeToString, ModuleInfo } from '@module-federation/sdk';
 import { getRemoteEntry } from '../utils/load';
 import { FederationHost } from '../core';
@@ -88,7 +88,7 @@ class Module {
         });
 
       if (typeof remoteEntryExports?.init === 'undefined') {
-        console.error(
+        logger.error(
           'The remote entry interface does not contain "init"',
           '\n',
           'Ensure the name of this remote is not reserved or in use. Check if anything already exists on window[nameOfRemote]',
@@ -114,8 +114,18 @@ class Module {
     this.lib = remoteEntryExports;
     this.inited = true;
 
+    let moduleFactory;
+    moduleFactory = await this.host.loaderHook.lifecycle.getModuleFactory.emit({
+      remoteEntryExports,
+      expose,
+      moduleInfo: this.remoteInfo,
+    });
+
     // get exposeGetter
-    const moduleFactory = await remoteEntryExports.get(expose);
+    if (!moduleFactory) {
+      moduleFactory = await remoteEntryExports.get(expose);
+    }
+
     assert(
       moduleFactory,
       `${getFMId(this.remoteInfo)} remote don't export ${expose}.`,