+++ packages/bridge/bridge-shared/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @module-federation/bridge-shared
 
+## 0.6.14
+
+### Patch Changes
+
+- ad605d2: chore: unified logger
+
+## 0.6.13
+
+## 0.6.12
+
 ## 0.6.11
 
 ## 0.6.10