+++ packages/data-prefetch/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @module-federation/data-prefetch
 
+## 0.6.14
+
+### Patch Changes
+
+- ad605d2: chore: unified logger
+- Updated dependencies [ad605d2]
+  - @module-federation/runtime@0.6.14
+  - @module-federation/sdk@0.6.14
+
+## 0.6.13
+
+### Patch Changes
+
+- @module-federation/runtime@0.6.13
+- @module-federation/sdk@0.6.13
+
+## 0.6.12
+
+### Patch Changes
+
+- @module-federation/runtime@0.6.12
+- @module-federation/sdk@0.6.12
+
 ## 0.6.11
 
 ### Patch Changes