+++ packages/rsbuild-plugin/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @module-federation/rsbuild-plugin
 
+## 0.6.14
+
+### Patch Changes
+
+- ad605d2: chore: unified logger
+- Updated dependencies [ad605d2]
+  - @module-federation/enhanced@0.6.14
+  - @module-federation/sdk@0.6.14
+
+## 0.6.13
+
+### Patch Changes
+
+- @module-federation/enhanced@0.6.13
+- @module-federation/sdk@0.6.13
+
+## 0.6.12
+
+### Patch Changes
+
+- cf14509: fix: configure the default assetPrefix for MF apps correctly
+  - @module-federation/enhanced@0.6.12
+  - @module-federation/sdk@0.6.12
+
 ## 0.0.2
 
 ### Patch Changes