+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,15 @@
 # [1.1.0-canary.1](https://github.com/module-federation/core/compare/sdk-1.0.0...sdk-1.1.0-canary.1) (2023-12-05)
 
+## 0.6.14
+
+### Patch Changes
+
+- ad605d2: chore: unified logger
+
+## 0.6.13
+
+## 0.6.12
+
 ## 0.6.11
 
 ### Patch Changes